requirements.txt
streamlit
pandas
tensorflow-cpu
scikit-learn
numpy